+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNKFE6XA2D6Q2C0TM7PC25PN"
+        "01KNKQ5S1TZMFFM5XEMJW0K48R"
     )
     r"""The unique identifier of the data source"""