+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNKFE6X0W3DHNMCPS548W3BM"
+        "01KNKQ5S1QDWQ955QXV0JST70V"
     )
     r"""The unique identifier of the data source"""