+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNKFE6XCN8EJCS36WVAHYWFM"
+        "01KNKQ5S1WYF9J7M3SR1204NDQ"
     )
     r"""The unique identifier of the data source"""